requirements.txt
python-telegram-bot[job-queue]
python-dotenv
google-generativeai
openai
litellm
